backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.7
langchain-community==0.3.7
langchain-groq==0.2.1
langchain-chroma==0.1.4
langgraph==0.2.53
chromadb==0.5.18
sentence-transformers==3.3.0
PyMuPDF==1.25.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
gtts==2.5.3
httpx==0.27.2
